Initial Commit
The code works right now! ...Except that it works too well. The nearest neighbor interpolation for the new pop2 grid cells interpolates directly from the 0s that were in place underneath the seafloor in the old restart file! Need to edit so that the interpolation applies to ONLY the predefined points in the old restart file.